mutation_predictor/requirements.txt
biopython>=1.81
pandas>=2.1.1
numpy>=1.26.4
scikit-learn>=1.3.1
matplotlib>=3.8.0
seaborn>=0.13.0

